Create #0:  Fourier's Seventeen Lines Problem
I lovingly call this program "problem 0." It truly is not a project Euler problem, or if it is, it is not indexed at 0. However, solving this problem lead me to discover project euler. At the end of last summer, I probably spent a entire afternoon staring at this program and figuring out what I wanted it to do. After finding the solution and gushing to my friends about it, one of them told me about Project Euler and started me off on that journey.